requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
openpyxl>=3.0.7
python-dotenv>=0.19.0